Updated Blend file and the texture screenshots
Textured the Main environment outside the museum. 
TODO:
Fix normals in Torus knot. There is something funky going on in the render.
Need to fix some topology errors to make our mountains look a little better. Also, Make them Lower poly? Yeah...
Model the roof of the museum.
Moden the outside of the museum walls so they can be textured.
Create a few models in the museum for a museum with stuff in it and not just boxy rooms.
Start to Scan in peoples Heads with the Kinect camera and get the topology cleaned up.